app/requirements.txt
django
django-redis
djangorestframework
channels
channels-redis
redis
psycopg2-binary
djangorestframework-simplejwt
Pillow